release stable v2.0.5

+++ src/node/filemanager.ts
@@ -1,13 +1,13 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-import Bluebird from 'bluebird';
-import findCacheDir from 'find-cache-dir';
 import * as fs from 'fs';
 import { default as nodePath } from 'path';
-import { cwd as nodeCwd } from 'process';
 import upath from 'upath';
-import { removeEmpties } from './array-utils';
-
+import ErrnoException = NodeJS.ErrnoException;
+import { cwd as nodeCwd } from 'process';
+import 'js-prototypes';
+import Bluebird from 'bluebird';
 import glob = require('glob');
+import findCacheDir from 'find-cache-dir';
 /**
  * node_modules/.cache/${name}
  */
@@ -22,32 +22,44 @@ const modPath = nodePath as Mutable<typeof nodePath>;
 /**
  * Directory iterator recursive
  * @param dir
- * @see {@link https://stackoverflow.com/a/66083078/6404439}
+ * @param done
  */
-function* walkSync(dir: fs.PathLike): Generator<string> {
-  const files = fs.readdirSync(dir, { withFileTypes: true });
-  for (const file of files) {
-    if (file.isDirectory()) {
-      yield* walkSync(join(dir, file.name));
-    } else {
-      yield join(dir, file.name);
-    }
-  }
-}
+// eslint-disable-next-line no-unused-vars
+const walk = function (dir: fs.PathLike, done: (err: ErrnoException | null, results?: string[]) => any) {
+  let results = [];
+  fs.readdir(dir, function (err, list) {
+    if (err) return done(err);
+    let pending = list.length;
+    if (!pending) return done(null, results);
+    list.forEach(function (file) {
+      file = modPath.resolve(dir.toString(), file);
+      fs.stat(file, function (err, stat) {
+        if (stat && stat.isDirectory()) {
+          walk(file, function (err, res) {
+            results = results.concat(res);
+            if (!--pending) done(null, results);
+          });
+        } else {
+          results.push(file);
+          if (!--pending) done(null, results);
+        }
+      });
+    });
+  });
+};
 
 const filemanager = {
-  /**
-   * @see {@link walkSync}
-   */
-  readdirSync: walkSync,
+  // eslint-disable-next-line no-unused-vars
+  readdirSync: (path: fs.PathLike, callback: (err: ErrnoException, results?: string[]) => any) => {
+    return walk(path, callback);
+  },
 
   /**
    * Remove dir or file recursive synchronously (non-empty folders supported)
    * @param path
    */
   rmdirSync: (path: fs.PathLike, options: fs.RmOptions = {}) => {
-    if (fs.existsSync(path))
-      return fs.rmSync(path, Object.assign({ recursive: true }, options));
+    if (fs.existsSync(path)) return fs.rmSync(path, Object.assign({ recursive: true }, options));
   },
 
   /**
@@ -57,11 +69,7 @@ const filemanager = {
    * @param callback
    * @returns
    */
-  rm: (
-    path: fs.PathLike,
-    options: fs.RmOptions | fs.NoParamCallback = {},
-    callback?: fs.NoParamCallback
-  ) => {
+  rm: (path: fs.PathLike, options: fs.RmOptions | fs.NoParamCallback = {}, callback?: fs.NoParamCallback) => {
     if (fs.existsSync(path)) {
       if (typeof options == 'function') {
         return fs.rm(path, { recursive: true }, options);
@@ -83,7 +91,6 @@ const filemanager = {
    * Write to file recursively (synchronous)
    * @param path
    * @param content
-   * @param append append to file?
    * @returns Promise.resolve(path);
    * @example
    * // write directly
@@ -93,21 +100,17 @@ const filemanager = {
    * // or log using async
    * input.then((file)=> console.log('written to', file));
    */
-  write: (path: fs.PathLike, content: any, append = false) => {
+  write: (path: fs.PathLike, content: any) => {
     const dir = modPath.dirname(path.toString());
     if (!fs.existsSync(dir)) fs.mkdirSync(dir, { recursive: true });
     if (typeof content != 'string') {
       if (typeof content == 'object') {
-        content = JSON.stringify(content, null, 4);
+        content = JSON.stringifyWithCircularRefs(content, 4);
       } else {
         content = String(content);
       }
     }
-    if (!append) {
-      fs.writeFileSync(path, content);
-    } else {
-      fs.appendFileSync(path, content);
-    }
+    fs.writeFileSync(path, content);
     return Bluebird.resolve(path);
   },
 
@@ -118,9 +121,8 @@ const filemanager = {
    * @returns
    */
   mkdirSync: (path: fs.PathLike, options: fs.MakeDirectoryOptions = {}) => {
-    if (!fs.existsSync(path))
-      return fs.mkdirSync(path, Object.assign({ recursive: true }, options));
-  }
+    if (!existsSync(path)) return fs.mkdirSync(path, Object.assign({ recursive: true }, options));
+  },
 };
 
 export function removeMultiSlashes(str: string) {
@@ -129,10 +131,7 @@ export function removeMultiSlashes(str: string) {
 
 export const globSrc = function (pattern: string, opts: glob.IOptions = {}) {
   return new Bluebird((resolve: (arg: string[]) => any, reject) => {
-    const opt: glob.IOptions = Object.assign(
-      { cwd: cwd(), dot: true, matchBase: true },
-      opts
-    );
+    const opt: glob.IOptions = Object.assign({ cwd: cwd(), dot: true, matchBase: true }, opts);
     glob(pattern, opt, function (err, files) {
       if (err) {
         return reject(err);
@@ -146,8 +145,7 @@ export default filemanager;
 export const normalize = upath.normalize;
 export const writeFileSync = filemanager.write;
 export const cwd = () => upath.toUnix(nodeCwd());
-export const dirname = (str: string) =>
-  removeMultiSlashes(upath.toUnix(upath.dirname(str)));
+export const dirname = (str: string) => removeMultiSlashes(upath.toUnix(upath.dirname(str)));
 interface ResolveOpt {
   [key: string]: any;
   /**
@@ -165,12 +163,12 @@ export const resolve = (str: string, opt: ResolveOpt | any = {}) => {
   const res = removeMultiSlashes(upath.toUnix(upath.resolve(str)));
   opt = Object.assign(
     {
-      validate: false
+      validate: false,
     },
     opt
   );
   if (opt.validate) {
-    if (fs.existsSync(res)) return res;
+    if (existsSync(res)) return res;
     return null;
   }
   return res;
@@ -181,16 +179,12 @@ export const resolve = (str: string, opt: ResolveOpt | any = {}) => {
  * @param opt
  * @returns
  */
-export function read(path: string): Buffer;
-export function read(path: string, opt?: BufferEncoding): string;
 export function read(
   path: string,
   opt?: Parameters<typeof fs.readFileSync>[1]
 ): ReturnType<typeof fs.readFileSync> | null {
-  if (fs.existsSync(path)) return fs.readFileSync(path, opt);
-  return null;
+  if (existsSync(path)) return readFileSync(path, opt);
 }
-
 /**
  * smart join to unix path
  * * removes empty/null/undefined
@@ -198,10 +192,11 @@ export function read(
  * @returns
  */
 export const join = (...str: any[]) => {
-  str = removeEmpties(str.map((s) => String(s)));
+  str = str.map((s) => String(s)).removeEmpties();
   return removeMultiSlashes(upath.toUnix(nodePath.join(...str)));
 };
 export const { write, readdirSync, rmdirSync, rm, mkdirSync } = filemanager;
 export const fsreadDirSync = fs.readdirSync;
+export const { existsSync, readFileSync, appendFileSync, statSync } = fs;
 export const { basename, relative, extname } = upath;
 export const PATH_SEPARATOR = modPath.sep;
